+++ CHANGELOG.md
@@ -14,6 +14,8 @@ FEATURES:
 
 * Add support for `group_by` and `secondary_rate` on resource `vault_quota_rate_limit`. Requires Vault Enterprise 1.20.0+ ([#2476](https://github.com/hashicorp/terraform-provider-vault/pull/2476))
 * Add support for Transit CMAC endpoint ([#2488](https://github.com/hashicorp/terraform-provider-vault/pull/2488))
+* Add support for AWS SSO profiles ([#1808](https://github.com/hashicorp/terraform-provider-vault/issues/1808))
+* Add support for role assumption in `auth_login_aws` for all credential sources ([#1672](https://github.com/hashicorp/terraform-provider-vault/issues/1672))
 
 * Add new resource `vault_scep_auth_backend_role` to manage roles in a SCEP auth backend. [#2479](https://github.com/hashicorp/terraform-provider-vault/pull/2479).
 * Add new datasource and resource `vault_pki_secret_backend_config_scep` for PKI SCEP configuration. [#2487](https://github.com/hashicorp/terraform-provider-vault/pull/2487).